add examples and reorganize the proto files
