+++ cmake/DawnOptions.cmake
@@ -30,6 +30,9 @@ option(DAWN_TESTING "Enable testing" ON)
 # Documentation
 option(DAWN_DOCUMENTATION "Enable documentation" OFF)
 
+# Build examples
+option(DAWN_EXAMPLES "Compile the examples" OFF)
+
 # Export options for meta projects
 yoda_export_options(DAWN 
   DAWN_ASSERTS
