+++ src/dawn/IIR/proto/IIR/IIR.proto
@@ -19,7 +19,7 @@
 syntax = "proto3";
 
 package dawn.proto.iir;
-import "statements.proto";
+import "SIR/statements.proto";
 
 /* ===-----------------------------------------------------------------------------------------===*/
 //      Accesses
