+++ src/dawn/SIR/CMakeLists.txt
@@ -18,7 +18,9 @@ include(DawnProtobufGenerate)
 dawn_protobuf_generate(
   OUT_FILES sir_proto_cpp_files
   OUT_INCLUDE_DIRS sir_proto_include_dirs
-  PROTOS ${CMAKE_CURRENT_SOURCE_DIR}/SIR.proto ${CMAKE_CURRENT_SOURCE_DIR}/statements.proto
+  WDIR ${CMAKE_CURRENT_SOURCE_DIR}/proto
+  PROTOS SIR/SIR.proto SIR/statements.proto
+  PACKG SIR
   LANGUAGE cpp
 )
 
@@ -39,7 +41,7 @@ yoda_add_library(
           ASTVisitor.h
           SIR.cpp
           SIR.h
-          SIR.proto
+	  proto/SIR/SIR.proto
           Statement.h
           ${sir_proto_cpp_files}
   OBJECT
@@ -55,14 +57,20 @@ yoda_combine_libraries(
 )
 
 # only include proto since protobuf uses relative path "statements.proto" -> export to dependent
-target_include_directories(DawnSIRStatic PUBLIC $<BUILD_INTERFACE:${sir_proto_include_dirs}> $<INSTALL_INTERFACE:include/>)
+target_include_directories(DawnSIRObjects PUBLIC ${sir_proto_include_dirs}/../)
+
+target_include_directories(DawnSIRStatic PUBLIC $<BUILD_INTERFACE:${sir_proto_include_dirs}/../> $<INSTALL_INTERFACE:include/>)
+target_include_directories(DawnSIRShared PUBLIC $<BUILD_INTERFACE:${sir_proto_include_dirs}/../> $<INSTALL_INTERFACE:include/>)
+
 
 add_custom_target(SIR_Proto_Generated ALL DEPENDS ${sir_proto_cpp_files})
 
 # Genreate python proto files
 dawn_protobuf_generate(
   OUT_FILES sir_proto_python_files
-  PROTOS ${CMAKE_CURRENT_SOURCE_DIR}/SIR.proto ${CMAKE_CURRENT_SOURCE_DIR}/statements.proto
+  PROTOS SIR/SIR.proto SIR/statements.proto
+  WDIR ${CMAKE_CURRENT_SOURCE_DIR}/proto
+  PACKG SIR
   LANGUAGE python
 )
 
@@ -75,7 +83,9 @@ add_custom_target(DawnSIRPythonProto ALL
 dawn_protobuf_generate(
   OUT_FILES sir_proto_java_files
   OUT_DIRS sir_proto_java_dirs
-  PROTOS ${CMAKE_CURRENT_SOURCE_DIR}/SIR.proto ${CMAKE_CURRENT_SOURCE_DIR}/statements.proto 
+  PROTOS SIR/SIR.proto SIR/statements.proto 
+  WDIR ${CMAKE_CURRENT_SOURCE_DIR}/proto
+  PACKG SIR
   LANGUAGE java
 )
 
@@ -84,5 +94,6 @@ add_custom_target(DawnSIRJavaProto ALL
   DEPENDS ${sir_proto_java_files}
 )
 
-install(FILES ${sir_proto_python_files} DESTINATION ${DAWN_INSTALL_PYTHON_DIR}/dawn)
+install(FILES ${sir_proto_python_files} DESTINATION ${DAWN_INSTALL_PYTHON_DIR}/SIR)
+install(FILES ${CMAKE_SOURCE_DIR}/python/SIR/__init__.py DESTINATION ${DAWN_INSTALL_PYTHON_DIR}/SIR)
 install(DIRECTORY ${sir_proto_java_dirs} DESTINATION ${DAWN_INSTALL_JAVA_DIR}/dawn)
