+++ src/dawn/SIR/proto/SIR/SIR.proto
@@ -19,7 +19,7 @@
 syntax = "proto3";
 
 package dawn.sir.proto;
-import "statements.proto";
+import "SIR/statements.proto";
 
 option java_package = "dawn.sir";
 option java_outer_classname = "SIR_pb2";
