+++ src/dawn/Serialization/ASTSerializer.h
@@ -17,7 +17,7 @@
 
 #include "dawn/IIR/StatementAccessesPair.h"
 #include "dawn/IIR/StencilInstantiation.h"
-#include "dawn/SIR/statements.pb.h"
+#include "dawn/SIR/SIR/statements.pb.h"
 #include "dawn/SIR/ASTVisitor.h"
 #include "dawn/SIR/SIR.h"
 #include <stack>
