+++ src/dawn/Serialization/SIRSerializer.cpp
@@ -15,7 +15,7 @@
 #include "dawn/SIR/SIR.h"
 #include "dawn/SIR/AST.h"
 #include "dawn/SIR/ASTVisitor.h"
-#include "dawn/SIR/SIR.pb.h"
+#include "dawn/SIR/SIR/SIR.pb.h"
 #include "dawn/Serialization/SIRSerializer.h"
 #include "dawn/Serialization/ASTSerializer.h"
 #include "dawn/Support/Format.h"
